requirements.txt
# Core ML/DL frameworks
tensorflow>=2.10.0,<2.16.0
numpy>=1.23.0,<2.0.0

# Image processing
Pillow>=9.0.0
albumentations>=1.3.0

# Data processing and utilities
scikit-learn>=1.2.0
matplotlib>=3.5.0
tqdm>=4.64.0

# Jupyter support (optional)
jupyter>=1.0.0
notebook>=6.5.0
ipykernel>=6.20.0
